backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
ffmpeg==1.4
numpy==1.22.3
opencv_python==4.7.0.72
tensorflow==2.10.0
